Use discovery subgraph also for planning, remove planner subgraph
