+++ packages/@n8n/ai-workflow-builder.ee/src/constants.ts
@@ -50,10 +50,7 @@ export const MAX_NODE_EXAMPLE_CHARS = 5000 * AVG_CHARS_PER_TOKEN_ANTHROPIC;
  */
 export const MAX_BUILDER_ITERATIONS = 30;
 export const MAX_CONFIGURATOR_ITERATIONS = 30;
+// Discovery handles both build mode and plan mode (planner merged into discovery)
 export const MAX_DISCOVERY_ITERATIONS = 50;
-export const MAX_PLANNER_ITERATIONS = 15;
 export const MAX_MULTI_AGENT_STREAM_ITERATIONS =
-	MAX_BUILDER_ITERATIONS +
-	MAX_CONFIGURATOR_ITERATIONS +
-	MAX_DISCOVERY_ITERATIONS +
-	MAX_PLANNER_ITERATIONS;
+	MAX_BUILDER_ITERATIONS + MAX_CONFIGURATOR_ITERATIONS + MAX_DISCOVERY_ITERATIONS;
