+++ packages/@n8n/ai-workflow-builder.ee/src/subgraphs/discovery.subgraph.ts
@@ -12,6 +12,8 @@ import { z } from 'zod';
 import { LLMServiceError } from '@/errors';
 import {
 	buildDiscoveryPrompt,
+	buildPlannerPrompt,
+	buildPlannerContextMessage,
 	formatTechniqueList,
 	formatExampleCategorizations,
 } from '@/prompts/agents/discovery.prompt';
@@ -28,8 +30,23 @@ import { createGetDocumentationTool } from '../tools/get-documentation.tool';
 import { createGetWorkflowExamplesTool } from '../tools/get-workflow-examples.tool';
 import { createNodeDetailsTool } from '../tools/node-details.tool';
 import { createNodeSearchTool } from '../tools/node-search.tool';
+import {
+	createSubmitQuestionsTool,
+	createSubmitPlanTool,
+	submitQuestionsSchema,
+	submitPlanSchema,
+	SUBMIT_QUESTIONS_TOOL,
+	SUBMIT_PLAN_TOOL,
+} from '../tools/planner-tools';
+import { extractConnectionChangingParameters } from '../tools/utils/connection.utils';
 import type { CoordinationLogEntry } from '../types/coordination';
-import { createDiscoveryMetadata } from '../types/coordination';
+import { createDiscoveryMetadata, createPlannerMetadata } from '../types/coordination';
+import type {
+	PlannerQuestion,
+	PlanOutput,
+	QuestionResponse,
+	PlannerInputMode,
+} from '../types/planner-types';
 import type { WorkflowMetadata } from '../types/tools';
 import { applySubgraphCacheMarkers } from '../utils/cache-control';
 import { buildWorkflowSummary, createContextMessage } from '../utils/context-builders';
@@ -69,20 +86,89 @@ const discoveryOutputSchema = z.object({
 });
 
 /**
+ * Question phase within the discovery/planner flow.
+ * Tracks where we are in the Q&A and planning flow.
+ */
+export type DiscoveryQuestionPhase =
+	| 'discovery' // Standard discovery mode (no questions)
+	| 'analyzing' // Analyzing the initial request (plan mode)
+	| 'asking' // Questions generated, waiting to emit (plan mode)
+	| 'planning' // Generating the plan (plan mode)
+	| 'complete'; // Done
+
+/**
  * Discovery Subgraph State
+ *
+ * Unified state for both discovery (build mode) and planning (plan mode).
+ * In build mode: discovers nodes without asking questions
+ * In plan mode: asks clarifying questions, then generates a plan with discovery context
  */
 export const DiscoverySubgraphState = Annotation.Root({
+	// ========================================================================
+	// Input fields
+	// ========================================================================
+
 	// Input: What the user wants to build
 	userRequest: Annotation<string>({
 		reducer: (x, y) => y ?? x,
 		default: () => '',
 	}),
 
+	// Input: Mode of operation
+	mode: Annotation<'build' | 'plan'>({
+		reducer: (x, y) => y ?? x,
+		default: () => 'build',
+	}),
+
+	// Input: Plan mode sub-mode (fresh, with_answers, refine)
+	plannerInputMode: Annotation<PlannerInputMode>({
+		reducer: (x, y) => y ?? x,
+		default: () => 'fresh',
+	}),
+
+	// Input: Existing workflow summary for context
+	existingWorkflowSummary: Annotation<string>({
+		reducer: (x, y) => y ?? x,
+		default: () => '',
+	}),
+
+	// Input: Existing plan for refinement mode
+	existingPlan: Annotation<PlanOutput | null>({
+		reducer: (x, y) => y ?? x,
+		default: () => null,
+	}),
+
+	// ========================================================================
+	// Internal state
+	// ========================================================================
+
 	// Internal: Conversation within this subgraph
 	messages: Annotation<BaseMessage[]>({
 		reducer: (x, y) => x.concat(y),
 		default: () => [],
 	}),
+
+	// Internal: Current phase in the question flow (plan mode)
+	questionPhase: Annotation<DiscoveryQuestionPhase>({
+		reducer: (x, y) => y ?? x,
+		default: () => 'discovery',
+	}),
+
+	// Internal: User's previous questions (for answer context)
+	previousQuestions: Annotation<PlannerQuestion[]>({
+		reducer: (x, y) => y ?? x,
+		default: () => [],
+	}),
+
+	// Internal: User's answers to questions (plan mode)
+	answers: Annotation<QuestionResponse[]>({
+		reducer: (x, y) => y ?? x,
+		default: () => [],
+	}),
+
+	// ========================================================================
+	// Output fields
+	// ========================================================================
 
 	// Output: Found nodes with version, reasoning, connection-changing parameters, and available resources
 	nodesFound: Annotation<
@@ -119,6 +205,28 @@ export const DiscoverySubgraphState = Annotation.Root({
 		default: () => [],
 	}),
 
+	// Output: Generated questions (plan mode)
+	questions: Annotation<PlannerQuestion[]>({
+		reducer: (x, y) => y ?? x,
+		default: () => [],
+	}),
+
+	// Output: Intro message before questions (plan mode)
+	introMessage: Annotation<string>({
+		reducer: (x, y) => y ?? x,
+		default: () => '',
+	}),
+
+	// Output: Generated plan (plan mode)
+	plan: Annotation<PlanOutput | null>({
+		reducer: (x, y) => y ?? x,
+		default: () => null,
+	}),
+
+	// ========================================================================
+	// Cached/passed-through fields
+	// ========================================================================
+
 	// Cached workflow templates (passed from parent, updated by tools)
 	cachedTemplates: Annotation<WorkflowMetadata[]>({
 		reducer: cachedTemplatesReducer,
@@ -136,6 +244,8 @@ export const DiscoverySubgraphState = Annotation.Root({
 export interface DiscoverySubgraphConfig {
 	parsedNodeTypes: INodeTypeDescription[];
 	llm: BaseChatModel;
+	/** Separate LLM for plan mode (defaults to llm if not provided) */
+	llmPlanner?: BaseChatModel;
 	logger?: Logger;
 	featureFlags?: BuilderFeatureFlags;
 }
@@ -146,9 +256,10 @@ export class DiscoverySubgraph extends BaseSubgraph<
 	typeof ParentGraphState.State
 > {
 	name = 'discovery_subgraph';
-	description = 'Discovers nodes and context for the workflow';
+	description = 'Discovers nodes and context for the workflow (unified with plan mode)';
 
-	private agent!: Runnable;
+	private discoveryAgent!: Runnable;
+	private plannerAgent!: Runnable;
 	private toolMap!: Map<string, StructuredTool>;
 	private logger?: Logger;
 	private parsedNodeTypes!: INodeTypeDescription[];
@@ -160,7 +271,7 @@ export class DiscoverySubgraph extends BaseSubgraph<
 		// Check if template examples are enabled
 		const includeExamples = config.featureFlags?.templateExamples === true;
 
-		// Create base tools
+		// Create base discovery tools (shared between discovery and plan mode)
 		const baseTools = [
 			createGetDocumentationTool(),
 			createNodeSearchTool(config.parsedNodeTypes),
@@ -172,20 +283,25 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			? [...baseTools, createGetWorkflowExamplesTool(config.logger)]
 			: baseTools;
 
+		// Build tool map for execution (discovery tools only - output tools don't need execution)
 		this.toolMap = new Map(tools.map((bt) => [bt.tool.name, bt.tool]));
 
-		// Define output tool
-		const submitTool = tool(() => {}, {
+		// ========================================================================
+		// Discovery Mode Agent (build mode - NEVER asks questions)
+		// ========================================================================
+
+		// Define discovery output tool
+		const submitDiscoveryTool = tool(() => {}, {
 			name: 'submit_discovery_results',
 			description: 'Submit the final discovery results',
 			schema: discoveryOutputSchema,
 		});
 
-		// Generate prompt based on feature flags
+		// Generate discovery prompt
 		const discoveryPrompt = buildDiscoveryPrompt({ includeExamples });
 
-		// Create agent with tools bound (including submit tool)
-		const systemPrompt = ChatPromptTemplate.fromMessages([
+		// Create discovery agent with tools bound
+		const discoverySystemPrompt = ChatPromptTemplate.fromMessages([
 			[
 				'system',
 				[
@@ -206,30 +322,79 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			});
 		}
 
-		// Bind all tools including the output tool
-		const allTools = [...tools.map((bt) => bt.tool), submitTool];
-		this.agent = systemPrompt.pipe(config.llm.bindTools(allTools));
+		// Discovery tools: discovery tools + submit_discovery_results
+		const discoveryAllTools = [...tools.map((bt) => bt.tool), submitDiscoveryTool];
+		this.discoveryAgent = discoverySystemPrompt.pipe(config.llm.bindTools(discoveryAllTools));
 
-		// Build the subgraph
+		// ========================================================================
+		// Planner Mode Agent (plan mode - SHOULD ask questions)
+		// ========================================================================
+
+		// Create planner-specific output tools
+		const submitQuestionsTool = createSubmitQuestionsTool();
+		const submitPlanTool = createSubmitPlanTool();
+
+		// Generate planner prompt
+		const plannerPrompt = buildPlannerPrompt({ includeExamples });
+
+		// Create planner agent with tools bound
+		const plannerSystemPrompt = ChatPromptTemplate.fromMessages([
+			[
+				'system',
+				[
+					{
+						type: 'text',
+						text: plannerPrompt,
+						cache_control: { type: 'ephemeral' },
+					},
+				],
+			],
+			['human', '{prompt}'],
+			['placeholder', '{messages}'],
+		]);
+
+		// Use separate LLM for planner if provided, otherwise fall back to discovery LLM
+		const plannerLLM = config.llmPlanner ?? config.llm;
+
+		if (typeof plannerLLM.bindTools !== 'function') {
+			throw new LLMServiceError('Planner LLM does not support tools', {
+				llmModel: plannerLLM._llmType(),
+			});
+		}
+
+		// Planner tools: discovery tools + submit_questions + submit_plan
+		const plannerAllTools = [...tools.map((bt) => bt.tool), submitQuestionsTool, submitPlanTool];
+		this.plannerAgent = plannerSystemPrompt.pipe(plannerLLM.bindTools(plannerAllTools));
+
+		// ========================================================================
+		// Build the unified subgraph
+		// ========================================================================
+
 		const subgraph = new StateGraph(DiscoverySubgraphState)
 			.addNode('agent', this.callAgent.bind(this))
 			.addNode('tools', async (state) => await executeSubgraphTools(state, this.toolMap))
 			.addNode('format_output', this.formatOutput.bind(this))
+			.addNode('format_questions', this.formatQuestions.bind(this))
+			.addNode('format_plan', this.formatPlan.bind(this))
 			.addEdge('__start__', 'agent')
-			// Conditional: tools if has tool calls, format_output if submit called
+			// Conditional routing based on which output tool was called
 			.addConditionalEdges('agent', this.shouldContinue.bind(this), {
 				tools: 'tools',
 				format_output: 'format_output',
-				end: END, // Fallback
+				format_questions: 'format_questions',
+				format_plan: 'format_plan',
+				end: END,
 			})
-			.addEdge('tools', 'agent') // After tools, go back to agent
-			.addEdge('format_output', END); // After formatting, END
+			.addEdge('tools', 'agent')
+			.addEdge('format_output', END)
+			.addEdge('format_questions', END) // HITL pause - questions generated
+			.addEdge('format_plan', 'format_output'); // Plan goes through format_output for hydration
 
 		return subgraph.compile();
 	}
 
 	/**
-	 * Agent node - calls discovery agent
+	 * Agent node - calls the appropriate agent based on mode
 	 * Context is already in messages from transformInput
 	 */
 	private async callAgent(state: typeof DiscoverySubgraphState.State) {
@@ -238,8 +403,27 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			applySubgraphCacheMarkers(state.messages);
 		}
 
-		// Messages already contain context from transformInput
-		const response = (await this.agent.invoke({
+		if (state.mode === 'plan') {
+			// Plan mode: Use planner agent with context message
+			const context = buildPlannerContextMessage({
+				userRequest: state.userRequest,
+				existingWorkflowSummary: state.existingWorkflowSummary || undefined,
+				previousPlan: state.existingPlan
+					? this.formatPlanForContext(state.existingPlan)
+					: undefined,
+				userAnswers: this.formatAnswersForContext(state.previousQuestions, state.answers),
+			});
+
+			const response = (await this.plannerAgent.invoke({
+				messages: state.messages,
+				prompt: context,
+			})) as AIMessage;
+
+			return { messages: [response] };
+		}
+
+		// Build mode: Use discovery agent with techniques
+		const response = (await this.discoveryAgent.invoke({
 			messages: state.messages,
 			prompt: state.userRequest,
 			techniques: formatTechniqueList(),
@@ -250,10 +434,16 @@ export class DiscoverySubgraph extends BaseSubgraph<
 	}
 
 	/**
-	 * Format the output from the submit tool call
+	 * Format the output from the submit tool call or plan hydration.
 	 * Hydrates availableResources for each node using node type definitions.
 	 */
 	private formatOutput(state: typeof DiscoverySubgraphState.State) {
+		// If we have a plan (from format_plan), hydrate its nodes
+		if (state.plan) {
+			return this.hydrateNodesFromPlan(state);
+		}
+
+		// Standard discovery output
 		const lastMessage = state.messages.at(-1);
 		let output: z.infer<typeof discoveryOutputSchema> | undefined;
 
@@ -298,6 +488,144 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			};
 		}
 
+		// Hydrate nodesFound with availableResources
+		const hydratedNodesFound = this.hydrateNodes(output.nodesFound, state.resourceOperationCache);
+
+		// Return hydrated output with best practices from state (updated by get_documentation tool)
+		return {
+			nodesFound: hydratedNodesFound,
+			bestPractices: state.bestPractices,
+			templateIds: state.templateIds ?? [],
+		};
+	}
+
+	/**
+	 * Format questions from the submit_questions tool call (plan mode)
+	 */
+	private formatQuestions(state: typeof DiscoverySubgraphState.State) {
+		const lastMessage = state.messages.at(-1);
+
+		if (!lastMessage || !isAIMessage(lastMessage) || !lastMessage.tool_calls) {
+			return {
+				questions: [],
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		const questionsCall = lastMessage.tool_calls.find(
+			(tc) => tc.name === SUBMIT_QUESTIONS_TOOL.toolName,
+		);
+
+		if (!questionsCall) {
+			return {
+				questions: [],
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		// Validate with schema
+		const parseResult = submitQuestionsSchema.safeParse(questionsCall.args);
+		if (!parseResult.success) {
+			this.logger?.error('[Discovery/Planner] Invalid questions schema', {
+				errors: parseResult.error.errors,
+			});
+			return {
+				questions: [],
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		const { questions, introMessage } = parseResult.data;
+
+		// Convert to PlannerQuestion format
+		const plannerQuestions: PlannerQuestion[] = questions.map((q) => ({
+			id: q.id,
+			question: q.question,
+			type: q.type,
+			options: q.options,
+			allowCustom: q.allowCustom ?? true,
+		}));
+
+		return {
+			questions: plannerQuestions,
+			introMessage: introMessage ?? '',
+			questionPhase: 'asking' as DiscoveryQuestionPhase,
+		};
+	}
+
+	/**
+	 * Format plan from the submit_plan tool call (plan mode)
+	 */
+	private formatPlan(state: typeof DiscoverySubgraphState.State) {
+		const lastMessage = state.messages.at(-1);
+
+		if (!lastMessage || !isAIMessage(lastMessage) || !lastMessage.tool_calls) {
+			return {
+				plan: null,
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		const planCall = lastMessage.tool_calls.find((tc) => tc.name === SUBMIT_PLAN_TOOL.toolName);
+
+		if (!planCall) {
+			return {
+				plan: null,
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		// Validate with schema
+		const parseResult = submitPlanSchema.safeParse(planCall.args);
+		if (!parseResult.success) {
+			this.logger?.error('[Discovery/Planner] Invalid plan schema', {
+				errors: parseResult.error.errors,
+			});
+			return {
+				plan: null,
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		const planData = parseResult.data;
+
+		// Build PlanOutput
+		const plan: PlanOutput = {
+			summary: planData.summary,
+			trigger: planData.trigger,
+			steps: planData.steps.map((s) => ({
+				description: s.description,
+				subSteps: s.subSteps,
+				suggestedNodes: s.suggestedNodes,
+			})),
+			additionalSpecs: planData.additionalSpecs,
+		};
+
+		return {
+			plan,
+			questionPhase: 'complete' as DiscoveryQuestionPhase,
+		};
+	}
+
+	// ========================================================================
+	// Hydration Helpers
+	// ========================================================================
+
+	/**
+	 * Hydrate nodes with availableResources from node type definitions or cache.
+	 */
+	private hydrateNodes(
+		nodes: Array<{
+			nodeName: string;
+			version: number;
+			reasoning: string;
+			connectionChangingParameters: Array<{
+				name: string;
+				possibleValues: Array<string | boolean | number>;
+			}>;
+		}>,
+		existingCache: Record<string, ResourceOperationInfo | null>,
+	) {
 		// Build lookup map for resource hydration
 		const nodeTypeMap = new Map<string, INodeTypeDescription>();
 		for (const nt of this.parsedNodeTypes) {
@@ -307,11 +635,8 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			}
 		}
 
-		// Get the resource operation cache from state
-		const existingCache = state.resourceOperationCache ?? {};
-
 		// Hydrate nodesFound with availableResources from node type definitions or cache
-		const hydratedNodesFound = output.nodesFound.map((node) => {
+		return nodes.map((node) => {
 			const cacheKey = createResourceCacheKey(node.nodeName, node.version);
 
 			// Check cache first (populated by node_details tool during discovery)
@@ -351,17 +676,149 @@ export class DiscoverySubgraph extends BaseSubgraph<
 				availableResources: resourceOpInfo.resources,
 			};
 		});
+	}
 
-		// Return hydrated output with best practices from state (updated by get_documentation tool)
+	/**
+	 * Hydrate nodes from plan suggestions.
+	 * Mirrors the hydration logic from the old planner.subgraph.ts buildDiscoveryContextFromPlan().
+	 */
+	private hydrateNodesFromPlan(state: typeof DiscoverySubgraphState.State) {
+		const plan = state.plan!;
+		const suggestedNodes = new Set<string>();
+
+		for (const step of plan.steps) {
+			if (step.suggestedNodes) {
+				step.suggestedNodes.forEach((n) => suggestedNodes.add(n));
+			}
+		}
+
+		if (suggestedNodes.size === 0) {
+			return {
+				nodesFound: [],
+				plan,
+				questionPhase: 'complete' as DiscoveryQuestionPhase,
+			};
+		}
+
+		// Build lookup maps for node types
+		// Plan may have display names ("HTTP Request") or internal names ("n8n-nodes-base.httpRequest")
+		const nodeTypeByInternalName = new Map<string, INodeTypeDescription>();
+		const nodeTypeByDisplayName = new Map<string, INodeTypeDescription>();
+
+		for (const nt of this.parsedNodeTypes) {
+			const ntMaxVersion = Array.isArray(nt.version) ? Math.max(...nt.version) : nt.version;
+
+			// Store by internal name - keep the one with highest version
+			const existingByName = nodeTypeByInternalName.get(nt.name);
+			if (!existingByName) {
+				nodeTypeByInternalName.set(nt.name, nt);
+			} else {
+				const existingMaxVersion = Array.isArray(existingByName.version)
+					? Math.max(...existingByName.version)
+					: existingByName.version;
+				if (ntMaxVersion > existingMaxVersion) {
+					nodeTypeByInternalName.set(nt.name, nt);
+				}
+			}
+
+			// Store by display name (e.g., "HTTP Request") - keep the one with highest version
+			const displayKey = nt.displayName.toLowerCase();
+			const existingByDisplay = nodeTypeByDisplayName.get(displayKey);
+			if (!existingByDisplay) {
+				nodeTypeByDisplayName.set(displayKey, nt);
+			} else {
+				const existingMaxVersion = Array.isArray(existingByDisplay.version)
+					? Math.max(...existingByDisplay.version)
+					: existingByDisplay.version;
+				if (ntMaxVersion > existingMaxVersion) {
+					nodeTypeByDisplayName.set(displayKey, nt);
+				}
+			}
+		}
+
+		// Debug: Log available node names for troubleshooting
+		this.logger?.debug('[Discovery/Planner] Hydration lookup maps built', {
+			internalNameCount: nodeTypeByInternalName.size,
+			displayNameCount: nodeTypeByDisplayName.size,
+			suggestedNodes: Array.from(suggestedNodes),
+		});
+
+		// Hydrate nodesFound with actual versions, available resources, and connection-changing parameters
+		const nodesFound = Array.from(suggestedNodes).map((nodeName) => {
+			// Try multiple matching strategies:
+			// 1. Exact internal name match (e.g., "n8n-nodes-base.httpRequest")
+			// 2. Exact display name match, case-insensitive (e.g., "HTTP Request")
+			// 3. Partial display name match (e.g., "Form Trigger" matches "n8n Form Trigger")
+			let nodeType =
+				nodeTypeByInternalName.get(nodeName) ?? nodeTypeByDisplayName.get(nodeName.toLowerCase());
+
+			// Try partial match if exact match fails
+			if (!nodeType) {
+				const nodeNameLower = nodeName.toLowerCase();
+				for (const [displayName, nt] of nodeTypeByDisplayName) {
+					// Check if suggested name contains the display name or vice versa
+					if (displayName.includes(nodeNameLower) || nodeNameLower.includes(displayName)) {
+						nodeType = nt;
+						this.logger?.debug('[Discovery/Planner] Partial match found', {
+							suggestedName: nodeName,
+							matchedDisplayName: displayName,
+						});
+						break;
+					}
+				}
+			}
+
+			if (!nodeType) {
+				// Log more details for debugging
+				this.logger?.warn('[Discovery/Planner] Node type not found during hydration', {
+					nodeName,
+					nodeNameLower: nodeName.toLowerCase(),
+					triedInternalName: nodeTypeByInternalName.has(nodeName),
+					triedDisplayName: nodeTypeByDisplayName.has(nodeName.toLowerCase()),
+				});
+				return {
+					nodeName,
+					version: 1,
+					reasoning: 'Suggested in plan step',
+					connectionChangingParameters: [] as Array<{
+						name: string;
+						possibleValues: Array<string | boolean | number>;
+					}>,
+				};
+			}
+
+			// Use the internal name for consistency with discovery output
+			const internalName = nodeType.name;
+
+			// Get the highest/latest version
+			const versions = Array.isArray(nodeType.version) ? nodeType.version : [nodeType.version];
+			const latestVersion = Math.max(...versions);
+
+			// Extract resource/operation info
+			const resourceOpInfo = extractResourceOperations(nodeType, latestVersion, this.logger);
+
+			// Extract connection-changing parameters from inputs/outputs expressions
+			const connectionChangingParameters = extractConnectionChangingParameters(nodeType);
+
+			return {
+				nodeName: internalName,
+				version: latestVersion,
+				reasoning: 'Suggested in plan step',
+				connectionChangingParameters,
+				...(resourceOpInfo && { availableResources: resourceOpInfo.resources }),
+			};
+		});
+
 		return {
-			nodesFound: hydratedNodesFound,
-			bestPractices: state.bestPractices,
-			templateIds: state.templateIds ?? [],
+			nodesFound,
+			plan,
+			questionPhase: 'complete' as DiscoveryQuestionPhase,
 		};
 	}
 
 	/**
 	 * Should continue with tools or finish?
+	 * Routes to appropriate format node based on which output tool was called.
 	 */
 	private shouldContinue(state: typeof DiscoverySubgraphState.State) {
 		const lastMessage = state.messages[state.messages.length - 1];
@@ -372,21 +829,39 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			lastMessage.tool_calls &&
 			lastMessage.tool_calls.length > 0
 		) {
-			// Check if the submit tool was called
+			// In plan mode, check for planner output tools first
+			if (state.mode === 'plan') {
+				// Check for submit_questions
+				const questionsCall = lastMessage.tool_calls.find(
+					(tc) => tc.name === SUBMIT_QUESTIONS_TOOL.toolName,
+				);
+				if (questionsCall) {
+					return 'format_questions';
+				}
+
+				// Check for submit_plan
+				const planCall = lastMessage.tool_calls.find((tc) => tc.name === SUBMIT_PLAN_TOOL.toolName);
+				if (planCall) {
+					return 'format_plan';
+				}
+			}
+
+			// Check for discovery submit tool
 			const submitCall = lastMessage.tool_calls.find(
 				(tc) => tc.name === 'submit_discovery_results',
 			);
 			if (submitCall) {
 				return 'format_output';
 			}
+
+			// Other tools - execute them
 			return 'tools';
 		}
 
 		// No tool calls = agent is done (or failed to call tool)
-		// In this pattern, we expect a tool call. If none, we might want to force it or just end.
-		// For now, let's treat it as an end, but ideally we'd reprompt.
 		this.logger?.warn(
-			'[Discovery] Agent stopped without calling submit_discovery_results - check if LLM is producing valid tool calls',
+			'[Discovery] Agent stopped without calling an output tool - check if LLM is producing valid tool calls',
+			{ mode: state.mode },
 		);
 		return 'end';
 	}
@@ -403,20 +878,54 @@ export class DiscoverySubgraph extends BaseSubgraph<
 		contextParts.push('</user_request>');
 
 		// 2. Current workflow summary (just node names, to know what exists)
-		// Discovery doesn't need full JSON, just awareness of existing nodes
+		let existingWorkflowSummary = '';
 		if (parentState.workflowJSON.nodes.length > 0) {
+			existingWorkflowSummary = buildWorkflowSummary(parentState.workflowJSON);
 			contextParts.push('<existing_workflow_summary>');
-			contextParts.push(buildWorkflowSummary(parentState.workflowJSON));
+			contextParts.push(existingWorkflowSummary);
 			contextParts.push('</existing_workflow_summary>');
 		}
 
 		// Create initial message with context
 		const contextMessage = createContextMessage(contextParts);
 
-		return {
+		// Base input (build mode)
+		const baseInput = {
 			userRequest,
-			messages: [contextMessage], // Context already in messages
+			messages: [contextMessage],
 			cachedTemplates: parentState.cachedTemplates,
+			existingWorkflowSummary,
+		};
+
+		// Build mode: simple discovery
+		if (parentState.mode !== 'plan') {
+			return {
+				...baseInput,
+				mode: 'build' as const,
+			};
+		}
+
+		// Plan mode: determine sub-mode (fresh, with_answers, refine)
+		let plannerInputMode: PlannerInputMode = 'fresh';
+		let answers: QuestionResponse[] = [];
+
+		// Check if this is an answers submission
+		const lastMessage = parentState.messages.at(-1);
+		if (lastMessage && this.isAnswersMessage(lastMessage)) {
+			plannerInputMode = 'with_answers';
+			answers = this.parseAnswersFromMessage(lastMessage);
+		} else if (parentState.planOutput && parentState.plannerPhase === 'plan_displayed') {
+			// Has existing plan and user sent new message = refinement
+			plannerInputMode = 'refine';
+		}
+
+		return {
+			...baseInput,
+			mode: 'plan' as const,
+			plannerInputMode,
+			answers,
+			previousQuestions: parentState.pendingQuestions ?? [],
+			existingPlan: parentState.planOutput ?? null,
 		};
 	}
 
@@ -431,26 +940,167 @@ export class DiscoverySubgraph extends BaseSubgraph<
 			bestPractices: subgraphOutput.bestPractices,
 		};
 
-		// Create coordination log entry (not a message)
+		// Base output
+		const baseOutput: Partial<typeof ParentGraphState.State> = {
+			discoveryContext,
+			templateIds,
+			cachedTemplates: subgraphOutput.cachedTemplates,
+		};
+
+		// Build mode: standard discovery output
+		if (subgraphOutput.mode !== 'plan') {
+			const logEntry: CoordinationLogEntry = {
+				phase: 'discovery',
+				status: 'completed',
+				timestamp: Date.now(),
+				summary: `Discovered ${nodesFound.length} nodes`,
+				metadata: createDiscoveryMetadata({
+					nodesFound: nodesFound.length,
+					nodeTypes: nodesFound.map((n) => n.nodeName),
+					hasBestPractices: !!subgraphOutput.bestPractices,
+				}),
+			};
+
+			return {
+				...baseOutput,
+				coordinationLog: [logEntry],
+			};
+		}
+
+		// Plan mode: include planner outputs
+		const { questions, answers, questionPhase, plan, introMessage } = subgraphOutput;
+
+		// Determine planner phase for parent state
+		let plannerPhase: 'idle' | 'waiting_for_answers' | 'plan_displayed' = 'idle';
+
+		if (questionPhase === 'asking' && questions.length > 0) {
+			plannerPhase = 'waiting_for_answers';
+		} else if (plan) {
+			plannerPhase = 'plan_displayed';
+		}
+
+		// Add discovery context to plan if generated
+		const planWithContext: PlanOutput | null = plan ? { ...plan, discoveryContext } : null;
+
+		// Create coordination log entry
 		const logEntry: CoordinationLogEntry = {
-			phase: 'discovery',
+			phase: 'planner',
 			status: 'completed',
 			timestamp: Date.now(),
-			summary: `Discovered ${nodesFound.length} nodes`,
-			metadata: createDiscoveryMetadata({
-				nodesFound: nodesFound.length,
-				nodeTypes: nodesFound.map((n) => n.nodeName),
-				hasBestPractices: !!subgraphOutput.bestPractices,
+			summary: plan
+				? `Generated plan with ${plan.steps.length} steps`
+				: `Generated ${questions.length} questions`,
+			metadata: createPlannerMetadata({
+				questionsAsked: questions.length,
+				questionsAnswered: answers.filter((a) => !a.skipped).length,
+				questionsSkipped: answers.filter((a) => a.skipped).length,
+				planGenerated: !!plan,
 			}),
 		};
 
 		return {
-			discoveryContext,
+			...baseOutput,
+			pendingQuestions: questions,
+			introMessage: introMessage ?? '',
+			planOutput: planWithContext,
+			plannerPhase,
 			coordinationLog: [logEntry],
-			// Pass template IDs for telemetry
-			templateIds,
-			// Propagate cached templates back to parent
-			cachedTemplates: subgraphOutput.cachedTemplates,
 		};
+	}
+
+	// ========================================================================
+	// Plan Mode Helper Methods
+	// ========================================================================
+
+	/**
+	 * Check if a message is an answers submission
+	 */
+	private isAnswersMessage(message: BaseMessage | undefined): boolean {
+		if (!message) return false;
+
+		// Check message content for question_answers type
+		const content = message.content;
+		if (typeof content === 'string') {
+			try {
+				const parsed = JSON.parse(content) as { type?: string };
+				return parsed.type === 'question_answers';
+			} catch {
+				return false;
+			}
+		}
+		return false;
+	}
+
+	/**
+	 * Parse answers from a message
+	 */
+	private parseAnswersFromMessage(message: BaseMessage): QuestionResponse[] {
+		const content = message.content;
+		if (typeof content === 'string') {
+			try {
+				const parsed = JSON.parse(content) as { answers?: QuestionResponse[] };
+				if (parsed.answers && Array.isArray(parsed.answers)) {
+					return parsed.answers;
+				}
+			} catch {
+				// Not JSON
+			}
+		}
+		return [];
+	}
+
+	/**
+	 * Format existing plan for context (used in refinement mode)
+	 */
+	private formatPlanForContext(plan: PlanOutput): string {
+		const parts: string[] = [];
+		parts.push(`Summary: ${plan.summary}`);
+		parts.push(`Trigger: ${plan.trigger}`);
+		parts.push('Steps:');
+		plan.steps.forEach((step, i) => {
+			parts.push(`${i + 1}. ${step.description}`);
+			if (step.subSteps?.length) {
+				step.subSteps.forEach((sub) => parts.push(`   - ${sub}`));
+			}
+		});
+		if (plan.additionalSpecs?.length) {
+			parts.push('Additional specs:');
+			plan.additionalSpecs.forEach((spec) => parts.push(`- ${spec}`));
+		}
+		return parts.join('\n');
+	}
+
+	/**
+	 * Format answers for context message (used when generating plan after Q&A)
+	 */
+	private formatAnswersForContext(
+		questions: PlannerQuestion[],
+		answers: QuestionResponse[],
+	): Array<{ question: string; answer: string }> | undefined {
+		if (!answers.length) return undefined;
+
+		const result: Array<{ question: string; answer: string }> = [];
+
+		for (const answer of answers) {
+			if (answer.skipped) continue;
+
+			const question = questions.find((q) => q.id === answer.questionId);
+			if (!question) continue;
+
+			let answerText = '';
+			if (answer.selectedOptions?.length) {
+				answerText = answer.selectedOptions.join(', ');
+			}
+			if (answer.customValue) {
+				answerText = answerText ? `${answerText}, ${answer.customValue}` : answer.customValue;
+			}
+
+			result.push({
+				question: question.question,
+				answer: answerText,
+			});
+		}
+
+		return result.length > 0 ? result : undefined;
 	}
 }
